Add missing refresh token when rotate and blacklist app are enabled
If ROTATE_REFRESH_TOKENS setting is set to True and blacklist app is enabled the refresh token that's sent when you call the token_refresh url is blacklisted.
The issue is that there is no new refresh token cookie sent back to the client.
This commit resolve that by creating and sending back a new refresh_token to the client.